+++ generators/app/templates/src/index.js
@@ -1,4 +1,5 @@
 import { App } from './components/App';
+import './style/style.scss';
 import ReactDOM from 'react-dom';
 import React from 'react';
 
@@ -6,12 +7,10 @@ window.showBack = dizmo.showBack;
 window.showFront = dizmo.showFront;
 
 document.addEventListener('dizmoready', () => {
-    ReactDOM.render(
-        <App />, document.getElementById('front')
-    );
-    document.getElementById('done').onclick = () => {
-        dizmo.showFront();
-    };
+    const front = document.getElementById('front');
+    ReactDOM.render(<App/>, front);
+    const done = document.getElementById('done');
+    done.onclick = () => dizmo.showFront();
 }, {
     once: true
 });